move the keybind handler to its own file and clean the code
